creation of the model_config.json, kubernetes support
