+++ kubernetes/scenescape-chart/templates/model-download/configmap.yaml
@@ -1,0 +1,21 @@
+# SPDX-FileCopyrightText: (C) 2026 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+{{- if and .Values.hooks.enabled .Values.modelDownload.enabled }}
+---
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: {{ .Release.Name }}-model-download
+  annotations:
+    "helm.sh/hook": pre-install
+    "helm.sh/hook-weight": "0"
+data:
+  download_models.py: |-
+{{ .Files.Get "files/model-download/download_models.py" | indent 4 }}
+  generate_model_config_from_models.py: |-
+{{ .Files.Get "files/model-download/generate_model_config.py" | indent 4 }}
+  models.json: |-
+{{ .Files.Get "files/model-download/models.json" | indent 4 }}
+{{- end }}
+
