+++ kubernetes/scenescape-chart/templates/model-installer/configmap.yaml
@@ -1,6 +1,7 @@
 # SPDX-FileCopyrightText: (C) 2025 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
+{{- if and .Values.hooks.enabled .Values.initModels.enabled (not .Values.modelDownload.enabled) }}
 ---
 apiVersion: v1
 kind: ConfigMap
@@ -22,3 +23,4 @@ data:
 {{ .Files.Get "files/model-installer/person-detection-retail-0013.json" | indent 4 }}
   model_config.json: |-
 {{ .Files.Get "files/model-installer/model_config.json" | indent 4 }}
+{{- end }}
