+++ model_download/README.md
@@ -14,13 +14,16 @@ models from various sources for purpose of demonstration.
 make -C model_download install-models
 ```
 
+For Kubernetes deployments, the same flow can be selected in the Helm chart with
+`--set modelDownload.enabled=true`. The chart runs the model downloader service and this folder's Python orchestration
+inside a pre-install hook, writing outputs to the models PVC.
+
 ## Key variables
 
 - `COMPOSE_PROJECT_NAME` (default: `scenescape`)
 - `MODEL_DOWNLOADER_IMAGE` (default: `intel/model-download:latest`)
 - `MODEL_DOWNLOADER_URL` (default: `http://127.0.0.1:8200`)
 - `MODEL_CONFIG_FILE` (default: `models.json`; shared model download and Scenescape config metadata)
-- `MODEL_LIST` (optional JSON array override; supports the format required by model_downloader REST API)
 - `MODEL_DOWNLOADER_CMD` (model_downloader startup arguments; default: `--plugins omz`)
 
 The downloader submits model download jobs, then polls `/api/v1/jobs` until every returned job reaches a terminal state.
@@ -28,7 +31,7 @@ The downloader submits model download jobs, then polls `/api/v1/jobs` until ever
 - Exit code `0`: all tracked download jobs completed successfully.
 - Exit code `1`: at least one tracked job failed, no job IDs were returned, polling timed out, or the status API returned an invalid response.
 
-After downloads finish, `make -C model_download install-models` also generates `model_config.json` in `/models/model_configs/` from `MODEL_CONFIG_FILE`.
+After downloads finish, `make -C model_download install-models` also generates `model_config.json` in `/models/model_configs/` from `MODEL_CONFIG_FILE`. If a model entry contains `scenescape.model_proc`, the same generation step writes that DL Streamer model-proc JSON file into the models volume and references it from generated `model_config.json`.
 
 ## Shared model configuration
 
@@ -50,10 +53,23 @@ The default configuration lives in `model_download/models.json`. Its top-level `
         "config": {
           "type": "detect",
           "params": {
-            "model_proc": "object_detection/person/person-detection-retail-0013.json"
+            "model": "omz/person-detection-retail-0013/FP16/person-detection-retail-0013.xml"
           },
           "adapter-params": {
             "metadatagenpolicy": "detectionPolicy"
+          }
+        },
+        "model_proc": {
+          "path": "object_detection/person/person-detection-retail-0013.json",
+          "content": {
+            "json_schema_version": "2.0.0",
+            "input_preproc": [],
+            "output_postproc": [
+              {
+                "labels": ["", "person"],
+                "converter": "tensor_to_bbox_ssd"
+              }
+            ]
           }
         }
       }
@@ -66,7 +82,12 @@ The default configuration lives in `model_download/models.json`. Its top-level `
 }
 ```
 
-`scenescape.name` is the convenient key written to generated `model_config.json`. If `scenescape.config.params.model` is omitted, the generator locates the downloaded `.xml` model under `/models/<hub>/<name>/` using the `model_downloader` section and prefers the precision configured by `model_config.prefer_precision`.
+`scenescape.name` is the convenient key written to generated `model_config.json`. `scenescape.config` is copied into the generated entry. `scenescape.config.params.model` must point to the model XML path relative to the models volume.
+
+`scenescape.model_proc` is optional. When present:
+- `path` is the model-proc JSON path relative to the models volume. It must be a relative `.json` path inside the models directory.
+- `content` is the DL Streamer model-proc JSON object to write at that path.
+- the generator adds `params.model_proc` with the same relative path to generated `model_config.json`.
 
 Models without the `scenescape` section are downloaded but skipped when generating `model_config.json`.
 
@@ -75,14 +96,6 @@ Use a different configuration file with:
 ```bash
 make -C model_download install-models MODEL_CONFIG_FILE=/path/to/models.json
 ```
-
-Example `MODEL_LIST` override for download-only experimentation:
-
-```bash
-MODEL_LIST='[{"name":"person-detection-retail-0013","hub":"omz"},{"name":"person-attributes-recognition-crossroad-0230","hub":"omz"}]' make -C model_download download-models
-```
-
-`MODEL_LIST` is a direct `model_downloader` payload override, so it does not use the nested `model_downloader` / `scenescape` structure.
 
 ## Downloading Models from Different Sources
 
@@ -95,7 +108,6 @@ To add another model source one has to:
 MODEL_DOWNLOADER_CMD ?= --plugins omz,huggingface
 ```
 
-- if, except of just downloading, the model needs some postprocessing - add this step in
-  the `Makefile` - in similar way to the `generate-model-config` or `copy-config-files` targets.
+- if the model needs DL Streamer post-processing metadata, add `scenescape.model_proc` to the model entry so the generator creates the model-proc file in the models volume.
 
 > [!NOTE] list of available plugins and their configuration can be found in the `model_downloader` [documentation](https://github.com/open-edge-platform/edge-ai-libraries/blob/main/microservices/model-download/README.md).
