--- model_download/src/copy-config-files
@@ -1,29 +1,0 @@
-#!/usr/bin/env python3
-
-# SPDX-FileCopyrightText: (C) 2025 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-
-from pathlib import Path
-import shutil
-import sys
-
-# map the source files inside scenescape source folder to their destination paths on models volume
-CONFIG_FILES = {
-  "dlstreamer-pipeline-server/model-proc-files/person-detection-retail-0013.json":"object_detection/person/person-detection-retail-0013.json"
-}
-
-def main():
-  if len(sys.argv) != 3:
-    print(f"Usage: {sys.argv[0]} <from_base> <to_base>")
-    sys.exit(1)
-  from_base, to_base = sys.argv[1], sys.argv[2]
-  from_base_path = Path(from_base)
-  to_base_path = Path(to_base)
-  for src_rel, dst_rel in CONFIG_FILES.items():
-    src_path = from_base_path / src_rel
-    dst_path = to_base_path / dst_rel
-    dst_path.parent.mkdir(parents=True, exist_ok=True)
-    shutil.copy2(src_path, dst_path)
-
-if __name__ == "__main__":
-  main()
